remove server-side rendering
