+++ index.html
@@ -5,10 +5,9 @@
     <link rel="icon" type="image/svg+xml" href="/vite.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Vite + Solid</title>
-    <!--app-head-->
   </head>
   <body>
-    <div id="root"><!--app-html--></div>
-    <script type="module" src="/src/entry-client.jsx"></script>
+    <div id="root"></div>
+    <script type="module" src="/src/client/main.tsx"></script>
   </body>
 </html>
