+++ src/client/main.tsx
@@ -1,0 +1,7 @@
+/* @refresh reload */
+import { render } from 'solid-js/web';
+
+import { App } from './app';
+import './index.css';
+
+render(() => <App />, document.getElementById('root') as HTMLElement);
