--- src/entry-server.tsx
@@ -1,11 +1,0 @@
-import { renderToString } from 'solid-js/web';
-
-import { App } from './client/app';
-
-/**
- * @param {string} _url
- */
-export function render(_url) {
-  const html = renderToString(() => <App />);
-  return { html };
-}
