+++ tsconfig.json
@@ -1,6 +1,9 @@
 {
   "include": ["src/**/*"],
   "compilerOptions": {
+    "target": "ES2020",
+    "module": "ESNext",
+    "moduleResolution": "Bundler",
     "noEmit": true,
     "skipLibCheck": true,
     "allowSyntheticDefaultImports": true,
