+++ vite.config.ts
@@ -1,7 +1,6 @@
 import { defineConfig } from 'vite';
 import solid from 'vite-plugin-solid';
 
-// https://vite.dev/config/
 export default defineConfig({
   plugins: [solid({ ssr: true })],
 });
